encore un joli commit sans nom et oui

+++ src/Controller/RecipeController.php
@@ -9,7 +9,10 @@ use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
+use Symfony\Component\HttpFoundation\JsonResponse;
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\HttpFoundation\File\Exception\FileException;
 
 #[Route('/recipe')]
 class RecipeController extends AbstractController
@@ -32,13 +35,36 @@ class RecipeController extends AbstractController
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
+            /** @var \App\Entity\User $user */
+            $user = $this->getUser();
+            if (!$user) {
+                throw $this->createAccessDeniedException('Vous devez être connecté pour créer une recette.');
+            }
+            $recipe->setUser($user);
             $recipe->setCreatedAt(new \DateTimeImmutable());
+
+            // Gestion de l'image uploadée
+            /** @var UploadedFile|null $imageFile */
+            $imageFile = $form->get('image')->getData();
+            if ($imageFile) {
+                $newFilename = uniqid().'_'.time().'.'.$imageFile->guessExtension();
+                try {
+                    $imageFile->move(
+                        $this->getParameter('recipes_directory'),
+                        $newFilename
+                    );
+                    $recipe->setImage($newFilename);
+                } catch (FileException $e) {
+                    $this->addFlash('danger', 'Erreur lors de l’upload de l’image.');
+                }
+            }
+
             $em->persist($recipe);
             $em->flush();
 
             $this->addFlash('success', 'Recette ajoutée !');
 
-            return $this->redirectToRoute('recipe_index');
+            return $this->redirectToRoute('profile');
         }
 
         return $this->render('recipe/new.html.twig', [
@@ -47,30 +73,44 @@ class RecipeController extends AbstractController
         ]);
     }
 
-    #[Route('/{id}', name: 'recipe_show', methods: ['GET'])]
-    public function show(Recipe $recipe): Response
-    {
-        return $this->render('recipe/show.html.twig', [
-            'recipe' => $recipe,
-        ]);
-    }
-
     #[Route('/{id}/edit', name: 'recipe_edit', methods: ['GET', 'POST'])]
     public function edit(Request $request, Recipe $recipe, EntityManagerInterface $em): Response
     {
+        /** @var \App\Entity\User $user */
+        $user = $this->getUser();
+        if (!$user || $recipe->getUser()->getId() !== $user->getId()) {
+            throw $this->createAccessDeniedException('Vous ne pouvez modifier que vos propres recettes.');
+        }
+
         $form = $this->createForm(RecipeType::class, $recipe);
         $form->handleRequest($request);
 
         if ($form->isSubmitted() && $form->isValid()) {
             $recipe->setUpdatedAt(new \DateTime());
-            $em->flush();
 
+            // Gestion de l'image uploadée
+            /** @var UploadedFile|null $imageFile */
+            $imageFile = $form->get('image')->getData();
+            if ($imageFile) {
+                $newFilename = uniqid().'_'.time().'.'.$imageFile->guessExtension();
+                try {
+                    $imageFile->move(
+                        $this->getParameter('recipes_directory'),
+                        $newFilename
+                    );
+                    $recipe->setImage($newFilename);
+                } catch (FileException $e) {
+                    $this->addFlash('danger', 'Erreur lors de l’upload de l’image.');
+                }
+            }
+
+            $em->flush();
             $this->addFlash('success', 'Recette modifiée !');
 
-            return $this->redirectToRoute('recipe_index');
+            return $this->redirectToRoute('profile');
         }
 
-        return $this->render('recipe/edit.html.twig', [
+        return $this->render('recipe/new.html.twig', [
             'recipe' => $recipe,
             'form' => $form,
         ]);
@@ -79,12 +119,30 @@ class RecipeController extends AbstractController
     #[Route('/{id}', name: 'recipe_delete', methods: ['POST'])]
     public function delete(Request $request, Recipe $recipe, EntityManagerInterface $em): Response
     {
+        /** @var \App\Entity\User $user */
+        $user = $this->getUser();
+        if (!$user || $recipe->getUser()->getId() !== $user->getId()) {
+            throw $this->createAccessDeniedException('Vous ne pouvez supprimer que vos propres recettes.');
+        }
+
         if ($this->isCsrfTokenValid('delete'.$recipe->getId(), $request->request->get('_token'))) {
             $em->remove($recipe);
             $em->flush();
             $this->addFlash('success', 'Recette supprimée !');
         }
 
-        return $this->redirectToRoute('recipe_index');
+        return $this->redirectToRoute('profile');
+    }
+
+    #[Route('/{id}/increment-view', name: 'recipe_increment_view', methods: ['POST'])]
+    public function incrementView(Recipe $recipe, EntityManagerInterface $em): JsonResponse
+    {
+        $recipe->setViews($recipe->getViews() + 1);
+        $em->flush();
+
+        return new JsonResponse([
+            'success' => true,
+            'views' => $recipe->getViews(),
+        ]);
     }
 }
